Move CacheRecord out to its own file.
